car and mensclothes only owner can detailed thier ads

+++ ad/views.py
@@ -11,6 +11,7 @@ from django.db import IntegrityError, ProgrammingError
 from rest_framework import generics
 from ad.filters import BagsKnapsacksFilter, CarFilter, CategoryFilter, ChildClothesShoesFilter, CustomFilterSet, CategoryFilterByName, MenClothesFilter, MenShoesFilter, WemenClothesFilter, WemenShoesFilter
 # from ad.func_for_help import choose_serializer
+from ad.func_for_help import add_view, check_if_authorised_has_profile
 from ad.models import BagsKnapsacks, Category, Car, ChildClothesShoes, Like, Images, MenClothes, MenShoes, Views, WemenClothes, WemenShoes
 from ad.pagination import OrdinaryListPagination
 from config.backends import CustomFilterQueryset, MyFilterBackend
@@ -225,23 +226,15 @@ class CarDetailGeneric(generics.RetrieveUpdateDestroyAPIView):
     )
     def get(self, request, pk, format=None):
         # Add new model instance Views get_or_creation
-        profile_instance = request.user.profile
+        # if request.user.is_anonymous:
+        #     return redirect(reverse("users:sign_in_email"))
+        # try:
+        #     profile_instance = request.user.profile
+        # except AttributeError as e:
+        #     print(e, 'Пользователь не авторизован, только просмотр объявления автомобиль')
         item = get_object_or_404(Car.objects.all(), pk=pk)
         serializer = CarPatchSerializer(item)
-        try:
-            profile = request.user.profile
-        except AttributeError:
-            print('---------------------', request.user, '4444444444', self.request.user)
-            return Response([serializer.data, {"message": "Anonymoususer, Views dont counted"}],
-                            status=status.HTTP_200_OK)
-        except Profile.DoesNotExist:
-            return Response({"message": "У вас нет Профиля, перенаправляем на регистрацию"},
-                            status=status.HTTP_203_NON_AUTHORITATIVE_INFORMATION)
-        # print('view created') #Here we create view every time we enter the object
-        # content_type = ContentType.objects.get(model='car').id
-        # view_instance = Views(profile=profile_instance, content_type=ContentType.objects.get_for_id(content_type), object_id=pk)
-        # view_instance.save()
-
+        add_view(serializer, request, pk)#Чтобы не повторять с 233 по 245 строчки вынесем функцию add_view
         return Response(serializer.data)
 
     @extend_schema(
@@ -252,7 +245,26 @@ class CarDetailGeneric(generics.RetrieveUpdateDestroyAPIView):
                     "включая все обязательные поля."
     )
     def put(self, request, *args, **kwargs):
-        return super().put(request, *args, **kwargs)  # Дла реализации доки
+        # if request.user.is_anonymous:
+        #     return  Response([{"message": "Anonymoususer, put method is not allowed"}],
+        #                 status=status.HTTP_405_METHOD_NOT_ALLOWED)
+        
+        # return super().put(request, *args, **kwargs)
+        pk=kwargs['pk']
+        check_if_authorised_has_profile(request)
+        # try:
+        #     profile = request.user.profile
+        # except AttributeError:
+        #     return Response([{"message": "Anonymoususer, put method is not allowed"}],
+        #                 status=status.HTTP_405_METHOD_NOT_ALLOWED)
+        # except Profile.DoesNotExist:
+        #     return Response({"message": "У вас нет Профиля, перенаправляем на регистрацию"},
+        #                 status=status.HTTP_203_NON_AUTHORITATIVE_INFORMATION)
+        if str(self.request.user.email)==str(Car.objects.get(id=pk).profilee.first()):
+            return super().put(request, *args, **kwargs)
+        else:
+            return Response({"message":"Хотя пользователь авторизирован, но объявление не его, редактировать, удалять нельзя"}, status=status.HTTP_401_UNAUTHORIZED)
+
 
     # def put(self, request, pk, format=None):
     #     item = get_object_or_404(Car.objects.all(), pk=pk)
@@ -269,36 +281,68 @@ class CarDetailGeneric(generics.RetrieveUpdateDestroyAPIView):
     )
   
     def patch(self, request, *args, **kwargs):
-        pk = kwargs['pk']
-        kwargs['partial'] = True 
-        print('--------------kwargs==', kwargs)
-        car_object = self.get_object()
-        serializer = CarPatchSerializer(car_object, data=request.data, partial=True) # set partial=True to update a data partially...CarUpdateImagesSerializer
+        pk=kwargs['pk']
+        check_if_authorised_has_profile(request)
+        # try:
+        #     profile = request.user.profile
+        # except AttributeError:
+        #     return Response([{"message": "Anonymoususer, put method is not allowed"}],
+        #                 status=status.HTTP_405_METHOD_NOT_ALLOWED)
+        # except Profile.DoesNotExist:
+        #     return Response({"message": "У вас нет Профиля, перенаправляем на регистрацию"},
+        #                 status=status.HTTP_203_NON_AUTHORITATIVE_INFORMATION)
+        car_object=self.get_object()
+        serializer=CarPatchSerializer(car_object, data=request.data, partial=True)
         if serializer.is_valid():
             serializer.save()
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
+            if str(self.request.user.email)==str(Car.objects.get(id=pk).profilee.first()):
+                return Response(serializer.data, status=status.HTTP_201_CREATED)
+            return Response({"message":"Хотя пользователь авторизирован, но объявление не его, редактировать, удалять нельзя"}, status=status.HTTP_401_UNAUTHORIZED)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+            # return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+            # return super().put(request, *args, **kwargs)
+        # else:
+        #     return Response([serializer.data, {"message":"Хотя пользователь авторизирован, но объявление не его, редактировать, удалять нельзя"}], status=status.HTTP_401_UNAUTHORIZED)
+
+
+
+        # if request.user.is_anonymous:
+        #     return  Response([{"message": "Anonymoususer, patch method is not allowed"}],
+        #                 status=status.HTTP_405_METHOD_NOT_ALLOWED)
+        # pk = kwargs['pk']
+        # kwargs['partial'] = True 
+        # print('--------------kwargs==', kwargs)
+        # car_object = self.get_object()
+        # serializer = CarPatchSerializer(car_object, data=request.data, partial=True) # set partial=True to update a data partially...CarUpdateImagesSerializer
+        # if serializer.is_valid():
+        #     serializer.save()
+        #     return Response(serializer.data, status=status.HTTP_201_CREATED)
+        # return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
     
     @extend_schema(
         methods=['DELETE'],
         summary="Удаление объекта"
     )
     def delete(self, request, *args, **kwargs):
-        pk=kwargs['pk'] 
+        pk=kwargs['pk']
+        check_if_authorised_has_profile(request)
         try:
             car_instanse = get_object_or_404(Car.objects.all(), pk=pk)
             print('car_instanse', car_instanse)
-        
-            images_instance = car_instanse.images.all()
-            if len(images_instance)>1:
-                for i in images_instance:
-                    i.delete
-                    i.save()
-            if len(images_instance)==1:
-                images_instance.delete()
+            if str(self.request.user.email)==str(Car.objects.get(id=pk).profilee.first()):
+                images_instance = car_instanse.images.all()
+                if len(images_instance)>1:
+                    for i in images_instance:
+                        i.delete
+                        i.save()
+                if len(images_instance)==1:
+                    images_instance.delete()
+                else:
+                    pass
+                car_instanse.delete()
             else:
-                pass
-            car_instanse.delete()
+                return Response({"message":"Хотя пользователь авторизирован, но объявление не его, редактировать, удалять нельзя"}, status=status.HTTP_401_UNAUTHORIZED)
+
         except Car.DoesNotExist as e:
                 print(e)
                 return Response({"message":"Theres no object with this id "})
